Разбиваю большие компоненты
