+++ project/src/components/card-film-screen/card-film-screen.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import {useHistory} from 'react-router-dom';
 import {AppRoute} from '../../const';
 import {Link} from 'react-router-dom';
@@ -19,7 +18,6 @@ type Props = {
 }
 
 function CardFilmScreen(props: Props): JSX.Element {
-  console.log('render');
   const {name, id, previewVideoLink, previewImage} = props;
   const history = useHistory();
   const [filmCardId, setFilmCardId] = useState(0);
